Write a line number to the whole field a class file gives it
A LineNumberTable entry's line_number is unsigned sixteen bits, per JVMS
4.7.12, but the entry was packed with addition and the line arrived as a
signed short. A line above Short.MAX_VALUE sign-extended and borrowed into
the start_pc half, leaving an entry the JVM rejects outright.

Mask instead, and take the line as an int. The short overload stays, since
ClassFileWriter is used outside Rhino, but nothing here calls it now.

+++ rhino/src/main/java/org/mozilla/javascript/optimizer/BodyCodegen.java
@@ -3247,13 +3247,13 @@ class BodyCodegen {
         int emitLine = remapLine(lineno, node.getColumn());
         if (emitLine == -1) return;
         itsLineNumber = emitLine;
-        cfw.addLineNumberEntry((short) emitLine);
+        cfw.addLineNumberEntry(emitLine);
     }
 
     private void addRemappedLineEntry(int line, int column) {
         int emitLine = remapLine(line, column);
         if (emitLine == -1) return;
-        cfw.addLineNumberEntry((short) emitLine);
+        cfw.addLineNumberEntry(emitLine);
     }
 
     private int remapLine(int line, int column) {
